requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-mock==3.12.0